requirements.txt
flask
flask-restful
flask-swagger-ui
flask-cors
requests
openai
mysql-connector-python